Modularized the ui a bit
