+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:circuitquest/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'ui/screens/home_screen.dart';
+import 'ui/home/home_screen.dart';
 import 'l10n/app_localizations.dart';
 import 'state/locale_provider.dart';
