+++ lib/ui/level_mode/difficulty_badge.dart
@@ -1,0 +1,63 @@
+import 'package:flutter/material.dart';
+
+/// Widget displaying the difficulty level with color coding.
+class DifficultyBadge extends StatelessWidget {
+  final String difficulty;
+
+  const DifficultyBadge({super.key, required this.difficulty});
+
+  @override
+  Widget build(BuildContext context) {
+    final color = _getDifficultyColor();
+    final icon = _getDifficultyIcon();
+
+    return Container(
+      decoration: BoxDecoration(
+        color: color.withAlpha((0.2 * 255).toInt()),
+        border: Border.all(color: color),
+        borderRadius: BorderRadius.circular(8),
+      ),
+      padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Icon(icon, size: 16, color: color),
+          const SizedBox(width: 8),
+          Text(
+            difficulty,
+            style: Theme.of(context).textTheme.labelSmall?.copyWith(
+                  color: color,
+                  fontWeight: FontWeight.bold,
+                ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Color _getDifficultyColor() {
+    switch (difficulty.toLowerCase()) {
+      case 'easy':
+        return Colors.green;
+      case 'medium':
+        return Colors.orange;
+      case 'hard':
+        return Colors.red;
+      default:
+        return Colors.grey;
+    }
+  }
+
+  IconData _getDifficultyIcon() {
+    switch (difficulty.toLowerCase()) {
+      case 'easy':
+        return Icons.sentiment_very_satisfied;
+      case 'medium':
+        return Icons.sentiment_neutral;
+      case 'hard':
+        return Icons.sentiment_very_dissatisfied;
+      default:
+        return Icons.help;
+    }
+  }
+}
