+++ lib/ui/shared/widgets/circuit_canvas/input_source_widget.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../../l10n/app_localizations.dart';
-import '../../core/components/input_source.dart';
-import '../../state/sandbox_state.dart';
-import '../utils/snackbar_utils.dart';
+import '../../../../l10n/app_localizations.dart';
+import '../../../../core/components/input_source.dart';
+import '../../../../state/sandbox_state.dart';
+import '../../utils/snackbar_utils.dart';
 
 /// A stateful widget to manage input source controls properly.
 class InputSourceWidget extends ConsumerStatefulWidget {
