+++ lib/ui/shared/widgets/control_panel.dart
@@ -1,12 +1,12 @@
 import 'package:circuitquest/l10n/app_localizations.dart';
-import 'package:circuitquest/ui/utils/snackbar_utils.dart';
+import 'package:circuitquest/ui/shared/utils/snackbar_utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../../state/sandbox_state.dart';
-import '../../state/level_state.dart';
-import '../../levels/levels.dart';
-import '../../core/components/input_source.dart';
-import '../../core/components/output_probe.dart';
+import '../../../state/sandbox_state.dart';
+import '../../../state/level_state.dart';
+import '../../../levels/levels.dart';
+import '../../../core/components/input_source.dart';
+import '../../../core/components/output_probe.dart';
 import 'circuit_file_manager.dart';
 
 /// Control panel for circuit simulation and evaluation.
